pyproject.toml
[tool.poetry]
name = "cal4"
version = "0.1.0"
description = ""
authors = ["Xerchio <xerchion@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
coverage = "^7.3.1"
Django = "^4.2.5"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
